int -> long long

+++ 11858.cpp
@@ -1,28 +1,28 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-typedef pair<int,int> ii;
+typedef pair<long long,long long> ii;
 typedef vector<ii> vii;
-typedef vector<int> vi;
+typedef vector<long long> vi;
 
-#define repr(i,l,r) for(int i=(l), _##i=(r); i<_##i; ++i)
+#define repr(i,l,r) for(long long i=(l), _##i=(r); i<_##i; ++i)
 #define d(arg) do {cerr << #arg << ": "; cerr << arg; cerr << endl;} while(0)
 #define pb push_back
-#define INF (int)1e9
+#define INF (long long)1e9
 #define EPS 1e-9
 
-pair<vi,int> inv(vi nums){
+pair<vi,long long> inv(vi nums){
     if(nums.size()==1)
         return make_pair(nums,0);
     vi left(nums.begin(),nums.begin()+nums.size()/2);
     vi right(nums.begin()+nums.size()/2,nums.end());
-    pair<vi,int> l = inv(left);
-    pair<vi,int> r = inv(right);
+    pair<vi,long long> l = inv(left);
+    pair<vi,long long> r = inv(right);
     vi larr=l.first, rarr=r.first;
-    int inversions = l.second+r.second;
+    long long inversions = l.second+r.second;
     vi ret;
-    int lp=0;
-    int rp=0;
+    long long lp=0;
+    long long rp=0;
     while(lp<larr.size()&&rp<rarr.size()){
         if(larr[lp]<=rarr[rp]){
             ret.pb(larr[lp]);
@@ -46,7 +46,7 @@ pair<vi,int> inv(vi nums){
 }
 
 int main(){
-    int N,i,j,k;
+    long long N,i,j,k;
     while(cin>>N){
         vi nums;
         repr(i,0,N){
